Starting to use Category instead of Category DTO
Also trying to balance not making an enormous change all at once, but CategoryDto is pervasive. Adding tests to the AddUpdateCategoryViewModel as well since that's getting touched.

Likely going to have Category get used just on the main screen for now, and leave the "leaf" screens to still use the DTO, to be refactored later

Finish AddUpdateCategoryViewModel cases

Add necessary ViewModel deps to Koin