Allow Phrase Text to Come from StringRes
This is part one of a strategy to make preset phrases all depend on resource IDs instead of strings, for easier runtime translation.

This commit adds in the infrastructure to the view/domain layer to be able to use resource IDs for phrases, but doesn't add any such phrases to the DB. This will be done in a later commit, when database migration/population are handled.