Refactor Category to not store the resource ID
Resource IDs are not stable. We were also needing to differentiate the type of category to get the category's name, this way that's abstracted away.
We still need another field to differentiate from the UI's idea of "hidden" and the idea of a preset category being permanently hidden. Until then, modifying the name of a preset category will result in it still appearing in settings.