Add notion of category pages
The idea is that this data can be observed by the page fragments, so that the pager adapter's content never has to actually change unless the page count changes. This should help us mitigate the display issues that are caused by the adapter rebinding any time the category list changes. This is not actually wired up to anything yet.